outline of what I'll be doing
thought a lot about how to travel back in time. The simplest thing I
came up with was to create a simple machine  and an assembly style
language. The machine is very similar to the interpreter provided but
with some extra state ( a program counter and a list of executed
instructions ).

keeping the original language DSL as it is much nicer than the assembly
- wrote a simple function to translate it for the machine.